Disambiguate 12 column grids
The popular 8 point grid system (https://builttoadapt.io/intro-to-the-8-point-grid-system-d2573cde8632) aligns on 8px width grid lines. Calling the vuetify grid system a "12 point grid" risks confusing it with a 12px base grid, when what we really mean is that it has 12 columns. To be clearer we should match the language used by other grid systems and call it a "12 column grid"